requirements.txt
moviepy>=2.0.0
Pillow>=10.0
requests>=2.31
openai>=1.0
pydub>=0.25
numpy>=1.24
